🎉 Initialize the repo with 0.0.1-alpha.1 draft
Start with somewhat self capable RFC draft and basic tests,
documentation material etc.

This is the first commit to follow Emotional commits standard.
Let that sink in.

(This will sound cool if people start using it in the future, but I am
fine with just entertaining myself when it goes the other way)

SemverImpact: patch
EmotionalCommits: 0.0.1-alpha.1